% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptivMC.R
\name{AdaptivMC}
\alias{AdaptivMC}
\title{Adaptive MC procedure}
\usage{
AdaptivMC(
  nOutput = 2,
  ndig = 2,
  p = 0.95,
  outliers = TRUE,
  FUN,
  n.iter = 10^4,
  ...
)
}
\arguments{
\item{nOutput}{\link{integer} (\strong{with default value}) number of model output(default nOutput=2)}

\item{ndig}{\link{integer} (\strong{with default value}) number of significant decimal digits (default ndig=2)}

\item{p}{\link{integer} (\strong{with default value}) coverage probability (default p= 0.95)}

\item{outliers}{\link{logical}(\strong{with default}) logical parameters indicating whether or not outliers are discarded. TRUE, the outliers are keeping, FALSE, the outliers are rejected ( default outliers=TRUE)}

\item{FUN}{\link{function} (\strong{required}) a simulation MC model}

\item{n.iter}{\link{integer} (\strong{with default}) number of iteration for each simulation (default n.iter= 1e4)}

\item{...}{further parameters passed to the FUN function}
}
\value{
a list including

ndig \link{integer} number of significant decimal digits

M \link{integer} the number of draws M= h x n.iter (h number of simulation and n.iter number of iteration for each simulation)

y \link{vector} an estimate of measurand Y

uy \link{vector} the standard uncertainties associated with the estimates,

Ry \link{matrix} correlation coefficients rij = r(yi, yj) associated with pairs of the estimates,

kp \link{integer} a coverage factor defining a 100p \% coverage region for Y,

Chain \link{mcmc} list of the simulations
}
\description{
adapted from GUM suppl.2
7.8 Adaptive Monte Carlo procedure
}
\examples{

\dontrun{
## example 9.2.2 GUM suppl.2 case1
Model<-
function(n.iter){
X1<-rnorm(n.iter,0,1)
X2<-rnorm(n.iter,0,1)
X3<-rnorm(n.iter,0,1)
Y1<-X1+X3
Y2<-X2+X3
cbind(Y1,Y2)
}

AdaptivMC(nOutput=2,ndig=3,p=0.95,FUN=Model,n.iter=1000)
}

\dontrun{
require(TLpack)

data(TLetru)
table<-Lum(TLetru,Doseb0=360,Dosea=0,alpha=FALSE,supra=FALSE)
B<-table$b
N<-table$n
table.norm<-B
 for (j in 1:3){
   for (i in 1:3)
     {
       table.norm[,i,j]<-(B[,i,j])/sum(N[seq(350,400),1,(j-1)*3+i])
       }
   }
 table.data<-cbind(table.norm[,,1],table.norm[,,2],table.norm[,,3])
 ii<-c(1,4,7,2,5,8,3,6,9)
 table.data<-table.data[seq(1,475),ii]
 Dose<-as.numeric(colnames(table.data))
 df.T<-matrix(rep(seq(26,500),9),475,9)
 df.y<-table.data[,1:9]
AdaptivMC(nOutput=5,ndig=2,outliers=FALSE,FUN=Slice5,Dose=Dose,df.T=df.T,df.y=df.y,n.thin=1)
}

}
\references{
JCGM-WG1 (2011) data – Supplement 2 to the “Guide to the expression of uncertainty in measurement” – Extension to any number of output quantities. Guide JCGM 102:2011. Sèvres: BIPM, IEC, IFCC, ILAC, ISO, IUPAC, IUPAP and OIML.
}
