% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gum.Output-package.R
\docType{package}
\name{Gum.Output-package}
\alias{Gum.Output}
\alias{Gum.Output-package}
\title{Gum.Output: Output following Gum supplement 1 (MCMC) and 2 (multivariate)}
\description{
Various function to analyze and plot OSL measurement using WinBUGS code.
}
\author{
\strong{Maintainer}: Antoine Zink \email{antoine.zink@culture.gouv.fr} (\href{https://orcid.org/0000-0001-7146-1101}{ORCID})

}
\keyword{internal}
