% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCM.Summary.R
\name{MCM.Summary}
\alias{MCM.Summary}
\title{MCM.Summary}
\usage{
MCM.Summary(output, p)
}
\arguments{
\item{output}{output quantity}

\item{p}{coverage probability

importFrom ("graphics", "contour")
importFrom ("stats", "sd")}
}
\value{
alist object with the following elements

$output.sort

$esp.sort

dev.sort

cov.map

map.p

low.cov

high.cov
}
\description{
final quantity presentation (output) MC GUM suppl2
}
