% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.Summary.R
\name{MCMC.Summary}
\alias{MCMC.Summary}
\title{MCMC.Summary}
\usage{
MCMC.Summary(output, p)
}
\arguments{
\item{output}{\link{matrix} (\strong{required}) output quantity}

\item{p}{\link{integer} (\strong{required}) coverage probability}
}
\value{
a list object with the following elements

$output.sort

$esp.sort

$dev.sort

$cov.map

$map.p

$low.cov low coverage value

$high.cov high coverage value
}
\description{
final quantity presentation (output) MC GUM supplement 2
}
