% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.Summary.R
\name{Output.Summary}
\alias{Output.Summary}
\title{Output.Summary}
\usage{
Output.Summary(output, p)
}
\arguments{
\item{output}{output quantity}

\item{p}{coverage probability}
}
\value{
a list object with the following elements

$output.sort

$esp.sort

dev.sort

cov.star

low.cov

high.cov
}
\description{
final quantity presentation (output) MC GUM suppl1
}
