% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveOutliers.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{RemoveOutliers}
\usage{
RemoveOutliers(tmp)
}
\arguments{
\item{tmp}{\link{matrix} (\strong{required}) matrix of MCMC}
}
\value{
a tmp matrix modified after discard of the outliers
}
\description{
procedure to remove outliers from matrix of MC simulation using the z-score
}
\details{
the function use the z-score.
}
\seealso{
outliers package especially grubbs.test (type=10) give the same result as z-score
}
