% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrait.R
\name{extrait}
\alias{extrait}
\title{extrait
#(changer le nom)}
\usage{
extrait(data, rang = 5, I1 = -200, I2 = 100)
}
\arguments{
\item{data}{\link[=mcmc object]{matrix} (\strong{required}) a matrix of MCMC output}

\item{rang}{\link{numeric} (\strong{with default value}) the column number}

\item{I1}{\link{numeric} (\strong{with default value}) the inferior limit}

\item{I2}{\link{numeric} (\strong{with default value}) the superior limit}
}
\value{
a matrix
}
\description{
#function additional to prepare file for Gum.Output package
#extract value between I1 and IZ in column n#rang
#equivalent to limit data using a uniform prior U\link{I1,I2} on variable rang
}
