% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.Summary.R
\name{Output.Summary}
\alias{Output.Summary}
\title{Output.Summary}
\usage{
Output.Summary(output, p)
}
\arguments{
\item{output}{output quantity}

\item{p}{coverage probability}
}
\value{
a list object with the following elements

$output.sort sorted output values (Gum suppl1 $7.5)

$esp.sort estimate of the output quantity (Gum suppl1 $7.6)

$dev.sort estimate of the associated standard uncertainty (Gum suppl1 $7.6)

$cov.star shortest coverage intervalle

$low.cov low coverage intervalle value

$high.cov high coverage intervalle value
}
\description{
final quantity presentation (output) MC GUM suppl1
}
